solve orboarding upload issue and delete orboarding image from aws

+++ src/app/modules/onboardingscreen/onboardingscreen.service.ts
@@ -7,6 +7,7 @@ import { IPaginationOptions } from '../../../interfaces/pagination';
 import { paginationHelper } from '../../../helpers/paginationHelper';
 import { onboardingscreenSearchableFields } from './onboardingscreen.constants';
 import { Types } from 'mongoose';
+import { S3Helper } from '../../../helpers/image/s3helper';
 
 
 const createOnboardingscreen = async (
@@ -16,7 +17,7 @@ const createOnboardingscreen = async (
   try {
     const result = await Onboardingscreen.create(payload);
     if (!result) {
-      
+
       throw new ApiError(
         StatusCodes.BAD_REQUEST,
         'Failed to create Onboardingscreen, please try again with valid data.'
@@ -25,7 +26,7 @@ const createOnboardingscreen = async (
 
     return result;
   } catch (error: any) {
-    
+
     if (error.code === 11000) {
       throw new ApiError(StatusCodes.CONFLICT, 'Duplicate entry found');
     }
@@ -107,6 +108,17 @@ const getSingleOnboardingscreen = async (id: string): Promise<IOnboardingscreen>
 const deleteOnboardingscreen = async (id: string): Promise<IOnboardingscreen> => {
   if (!Types.ObjectId.isValid(id)) {
     throw new ApiError(StatusCodes.BAD_REQUEST, 'Invalid Onboardingscreen ID');
+  }
+
+  const isExistOnboardingScreen = await Onboardingscreen.findById(id);
+  if (!isExistOnboardingScreen) {
+    throw new ApiError(StatusCodes.NOT_FOUND, 'Onboardingscreen not found');
+  }
+
+    if (isExistOnboardingScreen.imageURL) {
+    const url = new URL(isExistOnboardingScreen.imageURL)
+    const key = url.pathname.substring(1)
+    await S3Helper.deleteFromS3(key)
   }
 
   const result = await Onboardingscreen.findByIdAndDelete(id);
